refactor: to rename with-feature-toggle to branch-on-feature-toggle
