+++ packages/react-broadcast/modules/components/index.js
@@ -1,5 +1,5 @@
 export { default as FeatureToggled } from './feature-toggled';
 export { default as injectFeatureToggles } from './inject-feature-toggles';
 export { default as injectFeatureToggle } from './inject-feature-toggle';
-export { default as withFeatureToggle } from './with-feature-toggle';
+export { default as branchOnFeatureToggle } from './branch-on-feature-toggle';
 export { default as ConfigureFlopFlip } from './configure';
