+++ packages/react-redux/modules/index.js
@@ -15,7 +15,7 @@ export {
   FeatureToggled,
   injectFeatureToggle,
   injectFeatureToggles,
-  withFeatureToggle,
+  branchOnFeatureToggle,
   ConfigureFlopFlip,
 } from './components';
