+++ packages/react/modules/components/branch-on-feature-toggle.js
@@ -2,7 +2,7 @@ import { branch, renderNothing, renderComponent } from 'recompose';
 import isFeatureEnabled from '../helpers/is-feature-enabled';
 import { DEFAULT_FLAG_PROP_KEY } from '../constants';
 
-const branchUntoggled = (
+const branchOnFeatureToggle = (
   UntoggledComponent,
   flagName = DEFAULT_FLAG_PROP_KEY,
   flagVariate = true
@@ -12,4 +12,4 @@ const branchUntoggled = (
     UntoggledComponent ? renderComponent(UntoggledComponent) : renderNothing
   );
 
-export default branchUntoggled;
+export default branchOnFeatureToggle;
